refactor(analysis): split halstead token owners
