+++ crates/tokmd-analysis/src/halstead/mod.rs
@@ -9,453 +9,19 @@ use tokmd_analysis_types::{AnalysisLimits, normalize_path};
 
 const DEFAULT_MAX_FILE_BYTES: u64 = 128 * 1024;
 
-/// Languages that support Halstead analysis.
-pub(crate) fn is_halstead_lang(lang: &str) -> bool {
-    matches!(
-        lang.to_lowercase().as_str(),
-        "rust"
-            | "javascript"
-            | "typescript"
-            | "python"
-            | "go"
-            | "c"
-            | "c++"
-            | "java"
-            | "c#"
-            | "php"
-            | "ruby"
-    )
-}
+mod operators;
+mod tokenizer;
+
+pub(crate) use operators::is_halstead_lang;
+#[cfg(test)]
+pub(crate) use operators::operators_for_lang;
+#[cfg(test)]
+pub(crate) use tokenizer::FileTokenCounts;
+pub(crate) use tokenizer::tokenize_for_halstead;
 
 #[cfg(test)]
 #[path = "tests.rs"]
 mod moved_tests;
-
-/// Operators for a given language.
-pub(crate) fn operators_for_lang(lang: &str) -> &'static [&'static str] {
-    match lang.to_lowercase().as_str() {
-        "rust" => &[
-            "fn", "let", "mut", "if", "else", "match", "while", "for", "loop", "return", "break",
-            "continue", "struct", "enum", "impl", "trait", "pub", "use", "mod", "async", "await",
-            "move", "ref", "self", "Self", "where", "type", "const", "static", "unsafe", "extern",
-            "crate", "super", "as", "in", "->", "=>", "::", "=", "==", "!=", "<", ">", "<=", ">=",
-            "+", "-", "*", "/", "%", "&", "|", "^", "!", "&&", "||", "<<", ">>", "+=", "-=", "*=",
-            "/=", "%=", "&=", "|=", "^=", "<<=", ">>=", "..", "..=", "?",
-        ],
-        "javascript" | "typescript" => &[
-            "function",
-            "var",
-            "let",
-            "const",
-            "if",
-            "else",
-            "switch",
-            "case",
-            "while",
-            "for",
-            "do",
-            "return",
-            "break",
-            "continue",
-            "class",
-            "new",
-            "this",
-            "super",
-            "import",
-            "export",
-            "default",
-            "try",
-            "catch",
-            "finally",
-            "throw",
-            "async",
-            "await",
-            "yield",
-            "typeof",
-            "instanceof",
-            "in",
-            "of",
-            "delete",
-            "void",
-            "=>",
-            "=",
-            "==",
-            "===",
-            "!=",
-            "!==",
-            "<",
-            ">",
-            "<=",
-            ">=",
-            "+",
-            "-",
-            "*",
-            "/",
-            "%",
-            "&",
-            "|",
-            "^",
-            "!",
-            "&&",
-            "||",
-            "~",
-            "<<",
-            ">>",
-            ">>>",
-            "+=",
-            "-=",
-            "*=",
-            "/=",
-            "%=",
-            "&=",
-            "|=",
-            "^=",
-            "<<=",
-            ">>=",
-            ">>>=",
-            "**",
-            "**=",
-            "?.",
-            "??",
-            "??=",
-            "++",
-            "--",
-            "...",
-        ],
-        "python" => &[
-            "def", "class", "if", "elif", "else", "while", "for", "return", "break", "continue",
-            "import", "from", "as", "try", "except", "finally", "raise", "with", "yield", "lambda",
-            "pass", "del", "global", "nonlocal", "assert", "async", "await", "and", "or", "not",
-            "is", "in", "=", "==", "!=", "<", ">", "<=", ">=", "+", "-", "*", "/", "//", "%", "**",
-            "&", "|", "^", "~", "<<", ">>", "+=", "-=", "*=", "/=", "//=", "%=", "**=", "&=", "|=",
-            "^=", "<<=", ">>=", ":=",
-        ],
-        "go" => &[
-            "func",
-            "var",
-            "const",
-            "type",
-            "struct",
-            "interface",
-            "if",
-            "else",
-            "switch",
-            "case",
-            "for",
-            "range",
-            "return",
-            "break",
-            "continue",
-            "goto",
-            "defer",
-            "go",
-            "select",
-            "chan",
-            "map",
-            "package",
-            "import",
-            "fallthrough",
-            ":=",
-            "=",
-            "==",
-            "!=",
-            "<",
-            ">",
-            "<=",
-            ">=",
-            "+",
-            "-",
-            "*",
-            "/",
-            "%",
-            "&",
-            "|",
-            "^",
-            "!",
-            "&&",
-            "||",
-            "<<",
-            ">>",
-            "&^",
-            "+=",
-            "-=",
-            "*=",
-            "/=",
-            "%=",
-            "&=",
-            "|=",
-            "^=",
-            "<<=",
-            ">>=",
-            "&^=",
-            "++",
-            "--",
-            "<-",
-        ],
-        "c" | "c++" | "java" | "c#" | "php" => &[
-            "if",
-            "else",
-            "switch",
-            "case",
-            "while",
-            "for",
-            "do",
-            "return",
-            "break",
-            "continue",
-            "class",
-            "struct",
-            "new",
-            "delete",
-            "try",
-            "catch",
-            "throw",
-            "finally",
-            "void",
-            "static",
-            "const",
-            "public",
-            "private",
-            "protected",
-            "virtual",
-            "override",
-            "abstract",
-            "interface",
-            "extends",
-            "implements",
-            "import",
-            "include",
-            "using",
-            "namespace",
-            "typedef",
-            "template",
-            "=",
-            "==",
-            "!=",
-            "<",
-            ">",
-            "<=",
-            ">=",
-            "+",
-            "-",
-            "*",
-            "/",
-            "%",
-            "&",
-            "|",
-            "^",
-            "!",
-            "&&",
-            "||",
-            "~",
-            "<<",
-            ">>",
-            "+=",
-            "-=",
-            "*=",
-            "/=",
-            "%=",
-            "&=",
-            "|=",
-            "^=",
-            "<<=",
-            ">>=",
-            "++",
-            "--",
-            "->",
-            "::",
-            "sizeof",
-            "typeof",
-        ],
-        "ruby" => &[
-            "def",
-            "class",
-            "module",
-            "if",
-            "elsif",
-            "else",
-            "unless",
-            "while",
-            "until",
-            "for",
-            "do",
-            "return",
-            "break",
-            "next",
-            "begin",
-            "rescue",
-            "ensure",
-            "raise",
-            "yield",
-            "require",
-            "include",
-            "extend",
-            "attr_reader",
-            "attr_writer",
-            "attr_accessor",
-            "self",
-            "super",
-            "end",
-            "and",
-            "or",
-            "not",
-            "=",
-            "==",
-            "!=",
-            "<",
-            ">",
-            "<=",
-            ">=",
-            "+",
-            "-",
-            "*",
-            "/",
-            "%",
-            "&",
-            "|",
-            "^",
-            "!",
-            "&&",
-            "||",
-            "~",
-            "<<",
-            ">>",
-            "+=",
-            "-=",
-            "*=",
-            "/=",
-            "%=",
-            "&=",
-            "|=",
-            "^=",
-            "<<=",
-            ">>=",
-            "**",
-            "**=",
-            "<=>",
-            "=~",
-            "!~",
-            "..",
-            "...",
-        ],
-        _ => &[],
-    }
-}
-
-/// Per-file Halstead token counts.
-pub(crate) struct FileTokenCounts {
-    pub operators: BTreeMap<String, usize>,
-    pub operands: BTreeSet<String>,
-    pub total_operators: usize,
-    pub total_operands: usize,
-}
-
-/// Tokenize source code into operators and operands for Halstead analysis.
-pub(crate) fn tokenize_for_halstead(text: &str, lang: &str) -> FileTokenCounts {
-    let ops = operators_for_lang(lang);
-    let op_set: BTreeSet<&str> = ops.iter().copied().collect();
-
-    let mut operators: BTreeMap<String, usize> = BTreeMap::new();
-    let mut operands: BTreeSet<String> = BTreeSet::new();
-    let mut total_operators = 0usize;
-    let mut total_operands = 0usize;
-
-    for line in text.lines() {
-        let trimmed = line.trim();
-        // Skip comments and empty lines
-        if trimmed.is_empty()
-            || trimmed.starts_with("//")
-            || trimmed.starts_with('#')
-            || trimmed.starts_with("/*")
-            || trimmed.starts_with('*')
-        {
-            continue;
-        }
-
-        // Extract words and symbols
-        let mut chars = trimmed.chars().peekable();
-        while let Some(&ch) = chars.peek() {
-            if ch.is_whitespace() {
-                chars.next();
-                continue;
-            }
-
-            // Try to match multi-char operators (longest first)
-            if ch.is_ascii_punctuation() && ch != '_' && ch != '"' && ch != '\'' {
-                let remaining: String = chars.clone().take(4).collect();
-                let mut matched = false;
-                let char_count = remaining.chars().count();
-                for len in (1..=char_count.min(4)).rev() {
-                    let (byte_idx, _) = remaining
-                        .char_indices()
-                        .nth(len)
-                        .unwrap_or((remaining.len(), '\0'));
-                    let candidate = &remaining[..byte_idx];
-                    if op_set.contains(candidate) {
-                        *operators.entry(candidate.to_string()).or_insert(0) += 1;
-                        total_operators += 1;
-                        for _ in 0..len {
-                            chars.next();
-                        }
-                        matched = true;
-                        break;
-                    }
-                }
-                if !matched {
-                    chars.next();
-                }
-                continue;
-            }
-
-            // Identifiers/keywords
-            if ch.is_alphanumeric() || ch == '_' {
-                let mut word = String::new();
-                while let Some(&c) = chars.peek() {
-                    if c.is_alphanumeric() || c == '_' {
-                        word.push(c);
-                        chars.next();
-                    } else {
-                        break;
-                    }
-                }
-                if op_set.contains(word.as_str()) {
-                    *operators.entry(word).or_insert(0) += 1;
-                    total_operators += 1;
-                } else {
-                    operands.insert(word);
-                    total_operands += 1;
-                }
-                continue;
-            }
-
-            // String literals - skip
-            if ch == '"' || ch == '\'' {
-                chars.next();
-                let quote = ch;
-                while let Some(&c) = chars.peek() {
-                    chars.next();
-                    if c == '\\' {
-                        chars.next(); // skip escaped char
-                    } else if c == quote {
-                        break;
-                    }
-                }
-                total_operands += 1; // count the string literal as one operand
-                operands.insert("<string>".to_string());
-                continue;
-            }
-
-            chars.next();
-        }
-    }
-
-    FileTokenCounts {
-        operators,
-        operands,
-        total_operators,
-        total_operands,
-    }
-}
 
 /// Build aggregated Halstead metrics from source files.
 pub(crate) fn build_halstead_report(
