+++ crates/tokmd-analysis/src/halstead/operators.rs
@@ -1,0 +1,330 @@
+//! Language support and operator tables for Halstead analysis.
+
+/// Languages that support Halstead analysis.
+pub(crate) fn is_halstead_lang(lang: &str) -> bool {
+    matches!(
+        lang.to_lowercase().as_str(),
+        "rust"
+            | "javascript"
+            | "typescript"
+            | "python"
+            | "go"
+            | "c"
+            | "c++"
+            | "java"
+            | "c#"
+            | "php"
+            | "ruby"
+    )
+}
+
+/// Operators for a given language.
+pub(crate) fn operators_for_lang(lang: &str) -> &'static [&'static str] {
+    match lang.to_lowercase().as_str() {
+        "rust" => &[
+            "fn", "let", "mut", "if", "else", "match", "while", "for", "loop", "return", "break",
+            "continue", "struct", "enum", "impl", "trait", "pub", "use", "mod", "async", "await",
+            "move", "ref", "self", "Self", "where", "type", "const", "static", "unsafe", "extern",
+            "crate", "super", "as", "in", "->", "=>", "::", "=", "==", "!=", "<", ">", "<=", ">=",
+            "+", "-", "*", "/", "%", "&", "|", "^", "!", "&&", "||", "<<", ">>", "+=", "-=", "*=",
+            "/=", "%=", "&=", "|=", "^=", "<<=", ">>=", "..", "..=", "?",
+        ],
+        "javascript" | "typescript" => &[
+            "function",
+            "var",
+            "let",
+            "const",
+            "if",
+            "else",
+            "switch",
+            "case",
+            "while",
+            "for",
+            "do",
+            "return",
+            "break",
+            "continue",
+            "class",
+            "new",
+            "this",
+            "super",
+            "import",
+            "export",
+            "default",
+            "try",
+            "catch",
+            "finally",
+            "throw",
+            "async",
+            "await",
+            "yield",
+            "typeof",
+            "instanceof",
+            "in",
+            "of",
+            "delete",
+            "void",
+            "=>",
+            "=",
+            "==",
+            "===",
+            "!=",
+            "!==",
+            "<",
+            ">",
+            "<=",
+            ">=",
+            "+",
+            "-",
+            "*",
+            "/",
+            "%",
+            "&",
+            "|",
+            "^",
+            "!",
+            "&&",
+            "||",
+            "~",
+            "<<",
+            ">>",
+            ">>>",
+            "+=",
+            "-=",
+            "*=",
+            "/=",
+            "%=",
+            "&=",
+            "|=",
+            "^=",
+            "<<=",
+            ">>=",
+            ">>>=",
+            "**",
+            "**=",
+            "?.",
+            "??",
+            "??=",
+            "++",
+            "--",
+            "...",
+        ],
+        "python" => &[
+            "def", "class", "if", "elif", "else", "while", "for", "return", "break", "continue",
+            "import", "from", "as", "try", "except", "finally", "raise", "with", "yield", "lambda",
+            "pass", "del", "global", "nonlocal", "assert", "async", "await", "and", "or", "not",
+            "is", "in", "=", "==", "!=", "<", ">", "<=", ">=", "+", "-", "*", "/", "//", "%", "**",
+            "&", "|", "^", "~", "<<", ">>", "+=", "-=", "*=", "/=", "//=", "%=", "**=", "&=", "|=",
+            "^=", "<<=", ">>=", ":=",
+        ],
+        "go" => &[
+            "func",
+            "var",
+            "const",
+            "type",
+            "struct",
+            "interface",
+            "if",
+            "else",
+            "switch",
+            "case",
+            "for",
+            "range",
+            "return",
+            "break",
+            "continue",
+            "goto",
+            "defer",
+            "go",
+            "select",
+            "chan",
+            "map",
+            "package",
+            "import",
+            "fallthrough",
+            ":=",
+            "=",
+            "==",
+            "!=",
+            "<",
+            ">",
+            "<=",
+            ">=",
+            "+",
+            "-",
+            "*",
+            "/",
+            "%",
+            "&",
+            "|",
+            "^",
+            "!",
+            "&&",
+            "||",
+            "<<",
+            ">>",
+            "&^",
+            "+=",
+            "-=",
+            "*=",
+            "/=",
+            "%=",
+            "&=",
+            "|=",
+            "^=",
+            "<<=",
+            ">>=",
+            "&^=",
+            "++",
+            "--",
+            "<-",
+        ],
+        "c" | "c++" | "java" | "c#" | "php" => &[
+            "if",
+            "else",
+            "switch",
+            "case",
+            "while",
+            "for",
+            "do",
+            "return",
+            "break",
+            "continue",
+            "class",
+            "struct",
+            "new",
+            "delete",
+            "try",
+            "catch",
+            "throw",
+            "finally",
+            "void",
+            "static",
+            "const",
+            "public",
+            "private",
+            "protected",
+            "virtual",
+            "override",
+            "abstract",
+            "interface",
+            "extends",
+            "implements",
+            "import",
+            "include",
+            "using",
+            "namespace",
+            "typedef",
+            "template",
+            "=",
+            "==",
+            "!=",
+            "<",
+            ">",
+            "<=",
+            ">=",
+            "+",
+            "-",
+            "*",
+            "/",
+            "%",
+            "&",
+            "|",
+            "^",
+            "!",
+            "&&",
+            "||",
+            "~",
+            "<<",
+            ">>",
+            "+=",
+            "-=",
+            "*=",
+            "/=",
+            "%=",
+            "&=",
+            "|=",
+            "^=",
+            "<<=",
+            ">>=",
+            "++",
+            "--",
+            "->",
+            "::",
+            "sizeof",
+            "typeof",
+        ],
+        "ruby" => &[
+            "def",
+            "class",
+            "module",
+            "if",
+            "elsif",
+            "else",
+            "unless",
+            "while",
+            "until",
+            "for",
+            "do",
+            "return",
+            "break",
+            "next",
+            "begin",
+            "rescue",
+            "ensure",
+            "raise",
+            "yield",
+            "require",
+            "include",
+            "extend",
+            "attr_reader",
+            "attr_writer",
+            "attr_accessor",
+            "self",
+            "super",
+            "end",
+            "and",
+            "or",
+            "not",
+            "=",
+            "==",
+            "!=",
+            "<",
+            ">",
+            "<=",
+            ">=",
+            "+",
+            "-",
+            "*",
+            "/",
+            "%",
+            "&",
+            "|",
+            "^",
+            "!",
+            "&&",
+            "||",
+            "~",
+            "<<",
+            ">>",
+            "+=",
+            "-=",
+            "*=",
+            "/=",
+            "%=",
+            "&=",
+            "|=",
+            "^=",
+            "<<=",
+            ">>=",
+            "**",
+            "**=",
+            "<=>",
+            "=~",
+            "!~",
+            "..",
+            "...",
+        ],
+        _ => &[],
+    }
+}
